+++ Octokit.Tests.Integration/Clients/IssueCommentsClientTests.cs
@@ -219,7 +219,7 @@ public class IssueCommentsClientTests
             var numberToCreate = 2;
             using (var context = await _github.CreateRepositoryContextWithAutoInit(Helper.MakeNameWithTimestamp("IssueCommentsReactionTests")))
             {
-                var commentIds = new List<int>();
+                var commentIds = new List<long>();
 
                 // Create multiple test issues
                 for (int count = 1; count <= numberToCreate; count++)
@@ -404,7 +404,7 @@ public class IssueCommentsClientTests
             var numberToCreate = 2;
             using (var context = await _github.CreateRepositoryContextWithAutoInit(Helper.MakeNameWithTimestamp("IssueCommentsReactionTests")))
             {
-                var commentIds = new List<int>();
+                var commentIds = new List<long>();
 
                 // Create a single test issue
                 var issueNumber = await HelperCreateIssue(context.RepositoryOwner, context.RepositoryName);
@@ -594,7 +594,7 @@ public class IssueCommentsClientTests
         return issue.Number;
     }
 
-    async static Task<int> HelperCreateIssueCommentWithReactions(string owner, string repo, int number)
+    async static Task<long> HelperCreateIssueCommentWithReactions(string owner, string repo, int number)
     {
         var github = Helper.GetAuthenticatedClient();
